Fix code duplication in cpp_integration_test.rs
- Extract all duplicate test vectors to module-level constants
- Eliminate ~60 lines of repeated hardcoded test data
- Remove inappropriate console output from tests
- Use shared MESSAGE_HELLO constant across all tests
- Improve maintainability and follow DRY principles
